requirements.txt
python-telegram-bot==20.3
openai==1.0.0
flask==2.3.2
